DS-1984 Fix Specified Opening Times Smoke Test Bug (#976)
# Task Branch Pull Request

**<https://nhsd-jira.digital.nhs.uk/browse/DS-1984>**

## Description of Changes

This PR fixes a bug with the Smoke Test which the specified opening
times picks a date between now and the end of the year. As the year get
closer to ending such as December less dates can be picked without
picking a duplicate. So I've set the function to use `date_this_decade`
which reduces the chance of the same date being picked twice.