DS-2400 Integration Test and Unit Python package Bug (#1011)
# Task Branch Pull Request

**<https://nhsd-jira.digital.nhs.uk/browse/DS-2400>**

## Description of Changes

This fixes an issue with a python package after an update to a
dependency removed a underlying function. Fixed by removing pytest-bdd
package version pinning.

Also included in this PR is a temp fix for the build image codebuild
jobs. A 20 second sleep has been put in place to give time for the aws
ecr security image scan to be ready, before it's checked.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Development Checklist

- [x] I have performed a self-review of my own code
- [x] Tests have added that prove my fix is effective or that my feature
works (Integration tests)
- [x] I have updated Dependabot to include my changes (if applicable)

## Code Reviewer Checklist

- [x] I can confirm the changes have been tested or approved by a tester