DS-3268 Release pipeline xbuild bug fix (#1037)
# Task Branch Pull Request

**<https://nhsd-jira.digital.nhs.uk/browse/DS-3268>**

## Description of Changes
This PR updates the lambda image builder for pipelines to use an arm
based CodeBuild job. The 86x version stopped working due to an inability
to install the docker xbuild plugin. The arm builder we use for task
environments has not been affect by this issue. As such it makes sense
to align the non-pipeline builder with the pipeline builder and have
both CodeBuild jobs use an arm based container to build arm based
images.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Development Checklist

- [x] I have performed a self-review of my own code
- [x] Tests have added that prove my fix is effective or that my feature
works (Integration tests)
- [x] I have updated Dependabot to include my changes (if applicable)

## Code Reviewer Checklist

- [x] I can confirm the changes have been tested or approved by a tester